finals/SqueezeIn/requirements.txt
Flask==1.1.2
pyjwt==1.0.0
selenium==3.141.0

